refactor: extract shared page shell (DRY pass - signInPage, verifyPage, pageShell)
